// apps/mobile/src/features/mediaStore/mediaStore.types.ts
export type NativeAudioTrack = {
  id: string;
  title?: string;
  artist?: string;
  album?: string;
  albumId?: string;
  url: string;
  contentUri: string;
  duration?: number;
  mimeType?: string;
  size?: number;
  fileName?: string;
  folderPath?: string;
  fileSize?: number;
  dateAdded?: number;
  dateModified?: number;
  trackNumber?: number;
  localArtworkUri?: string;
};